Moving api files to their own directory in td.vue
