+++ td.vue/src/views/OauthReturn.vue
@@ -7,7 +7,7 @@
 import { mapState } from 'vuex';
 
 import { AUTH_SET_JWT } from '@/store/actions/auth.js';
-import loginApi from '@/service/loginApi.js';
+import loginApi from '@/service/api/loginApi.js';
 
 export default {
     name: 'OAuthReturn',
