+++ td.vue/tests/unit/views/home.spec.js
@@ -5,7 +5,7 @@ import Vuex from 'vuex';
 
 import { AUTH_SET_LOCAL } from '@/store/actions/auth.js';
 import Home from '@/views/Home.vue';
-import loginApi from '@/service/loginApi.js';
+import loginApi from '@/service/api/loginApi.js';
 import { PROVIDER_SELECTED } from '@/store/actions/provider.js';
 import router from '@/router/index.js';
 import TdProviderLoginButton from '@/components/ProviderLoginButton.vue';
